Keycloak migration

+++ src/main/java/com/keepitup/magjobbackend/assignee/controller/impl/AssigneeDefaultController.java
@@ -9,6 +9,9 @@ import com.keepitup.magjobbackend.assignee.function.AssigneeToResponseFunction;
 import com.keepitup.magjobbackend.assignee.function.AssigneesToResponseFunction;
 import com.keepitup.magjobbackend.assignee.function.RequestToAssigneeFunction;
 import com.keepitup.magjobbackend.assignee.service.api.AssigneeService;
+import com.keepitup.magjobbackend.configuration.Constants;
+import com.keepitup.magjobbackend.configuration.SecurityService;
+import com.keepitup.magjobbackend.organization.entity.Organization;
 import com.keepitup.magjobbackend.task.entity.Task;
 import com.keepitup.magjobbackend.task.service.api.TaskService;
 import lombok.extern.java.Log;
@@ -30,24 +33,35 @@ public class AssigneeDefaultController implements AssigneeController {
     private final AssigneeToResponseFunction assigneeToResponse;
     private final AssigneesToResponseFunction assigneesToResponse;
     private final TaskService taskService;
+    private final SecurityService securityService;
 
     public AssigneeDefaultController(
             AssigneeService service,
             RequestToAssigneeFunction requestToAssignee,
             AssigneeToResponseFunction assigneeToResponse,
             AssigneesToResponseFunction assigneesToResponse,
-            TaskService taskService
+            TaskService taskService,
+            SecurityService securityService
     ) {
         this.service = service;
         this.requestToAssignee = requestToAssignee;
         this.assigneeToResponse = assigneeToResponse;
         this.assigneesToResponse = assigneesToResponse;
         this.taskService = taskService;
+        this.securityService = securityService;
     }
 
 
     @Override
     public GetAssigneeResponse getAssignee(BigInteger memberId, BigInteger taskId) {
+        Organization organization = taskService.find(taskId).orElseThrow(
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND)
+        ).getOrganization();
+
+        if(!securityService.belongsToOrganization(organization)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         return service.findByMemberAndTask(memberId, taskId)
                 .map(assigneeToResponse)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
@@ -55,6 +69,15 @@ public class AssigneeDefaultController implements AssigneeController {
 
     @Override
     public GetAssigneeResponse assignMemberToTask(PostAssigneeRequest request) {
+        Organization organization = taskService.find(request.getTask()).orElseThrow(
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND)
+        ).getOrganization();
+
+        if (!securityService.hasPermission(organization, Constants.PERMISSION_NAME_CAN_MANAGE_TASKS)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
+
         Optional<Assignee> assignee = service.findByMemberAndTask(request.getMember(), request.getTask());
 
         if (assignee.isPresent()) {
@@ -70,6 +93,14 @@ public class AssigneeDefaultController implements AssigneeController {
 
     @Override
     public GetAssigneesResponse getAssigneesByTask(int page, int size, BigInteger taskId) {
+        Organization organization = taskService.find(taskId).orElseThrow(
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND)
+        ).getOrganization();
+
+        if(!securityService.belongsToOrganization(organization)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+  
         PageRequest pageRequest = PageRequest.of(page, size);
         Optional<Task> task = taskService.find(taskId);
 
@@ -82,6 +113,14 @@ public class AssigneeDefaultController implements AssigneeController {
 
     @Override
     public void deleteMemberFromTask(PostAssigneeRequest request) {
+        Organization organization = taskService.find(request.getTask()).orElseThrow(
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND)
+        ).getOrganization();
+
+        if (!securityService.hasPermission(organization, Constants.PERMISSION_NAME_CAN_MANAGE_TASKS)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         service.findByMemberAndTask(request.getMember(), request.getTask())
                 .ifPresentOrElse(
                         assignee -> service.delete(request.getMember(), request.getTask()),
