+++ src/main/java/com/keepitup/magjobbackend/user/controller/impl/UserDefaultController.java
@@ -1,61 +1,50 @@
 package com.keepitup.magjobbackend.user.controller.impl;
 
+import com.keepitup.magjobbackend.configuration.SecurityService;
+import com.keepitup.magjobbackend.jwt.CustomJwt;
 import com.keepitup.magjobbackend.user.controller.api.UserController;
 import com.keepitup.magjobbackend.user.dto.*;
 import com.keepitup.magjobbackend.user.entity.User;
 import com.keepitup.magjobbackend.user.function.*;
-import com.keepitup.magjobbackend.user.service.impl.AuthenticationUserService;
 import com.keepitup.magjobbackend.user.service.impl.UserDefaultService;
-import com.keepitup.magjobbackend.user.utils.JwtUtil;
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.java.Log;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.DisabledException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.io.IOException;
-import java.math.BigInteger;
 import java.util.Optional;
+import java.util.UUID;
 
 @RestController
 @Log
 public class UserDefaultController implements UserController {
     private final UserDefaultService service;
-    private final AuthenticationUserService authenticationService;
     private final UserToResponseFunction userToResponse;
     private final UsersToResponseFunction usersToResponse;
     private final RequestToUserFunction requestToUser;
     private final UpdateUserWithRequestFunction updateUserWithRequest;
-    private final UpdateUserPasswordWithRequestFunction updateUserPasswordWithRequestFunction;
-    private final AuthenticationManager authenticationManager;
-    private final JwtUtil jwtUtil;
+    private final SecurityService securityService;
 
     @Autowired
     public UserDefaultController(
             UserDefaultService service,
-            AuthenticationUserService authenticationService, UserToResponseFunction userToResponse,
+            UserToResponseFunction userToResponse,
             UsersToResponseFunction usersToResponse,
             RequestToUserFunction requestToUser,
             UpdateUserWithRequestFunction updateUserWithRequest,
             UpdateUserPasswordWithRequestFunction updateUserPasswordWithRequestFunction,
-            AuthenticationManager authenticationManager, JwtUtil jwtUtil) {
+            SecurityService securityService
+    ) {
         this.service = service;
-        this.authenticationService = authenticationService;
         this.userToResponse = userToResponse;
         this.usersToResponse = usersToResponse;
         this.requestToUser = requestToUser;
         this.updateUserWithRequest = updateUserWithRequest;
-        this.updateUserPasswordWithRequestFunction = updateUserPasswordWithRequestFunction;
-        this.authenticationManager = authenticationManager;
-        this.jwtUtil = jwtUtil;
+        this.securityService = securityService;
     }
 
     @Override
@@ -66,28 +55,51 @@ public class UserDefaultController implements UserController {
     }
 
     @Override
-    public GetUserResponse getUser(BigInteger id) {
+    @PreAuthorize("isAuthenticated()")
+    public GetUserResponse getUser(UUID id) {
+        var jwt = (CustomJwt) SecurityContextHolder.getContext().getAuthentication();
+        UUID loggedInUserId = UUID.fromString(jwt.getExternalId());
+
+        if (!loggedInUserId.equals(id)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not authorized to access this profile.");
+        }
+
         return service.find(id)
                 .map(userToResponse)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
     }
 
     @Override
-    public GetUserResponse createUser(PostUserRequest postUserRequest) {
-        Optional<User> user = service.find(postUserRequest.getEmail());
+    @PreAuthorize("permitAll()")
+    public GetUserResponse createUser() {
+        var jwt = (CustomJwt) SecurityContextHolder.getContext().getAuthentication();
+
+        Optional<User> user = service.find(UUID.fromString(jwt.getExternalId()));
+
+        PostUserRequest postUserRequest;
         if (user.isPresent()) {
             throw new ResponseStatusException(HttpStatus.CONFLICT);
-        }
-        else {
+        } else {
+            postUserRequest = new PostUserRequest();
+            postUserRequest.setEmail(jwt.getEmail());
+            postUserRequest.setFirstname(jwt.getFirstname());
+            postUserRequest.setLastname(jwt.getLastname());
+            postUserRequest.setId(UUID.fromString(jwt.getExternalId()));
+
             service.register(requestToUser.apply(postUserRequest));
         }
+
         return service.find(postUserRequest.getEmail())
                 .map(userToResponse)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
     }
 
     @Override
-    public void deleteUser(BigInteger id) {
+    public void deleteUser(UUID id) {
+        if (!securityService.hasAdminPermission()) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         service.find(id)
                 .ifPresentOrElse(
                         user -> service.delete(id),
@@ -98,7 +110,14 @@ public class UserDefaultController implements UserController {
     }
 
     @Override
-    public GetUserResponse updateUser(BigInteger id, PatchUserRequest patchUserRequest) {
+    public GetUserResponse updateUser(UUID id, PatchUserRequest patchUserRequest) {
+        var jwt = (CustomJwt) SecurityContextHolder.getContext().getAuthentication();
+        UUID loggedInUserId = UUID.fromString(jwt.getExternalId());
+
+        if (!loggedInUserId.equals(id)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not authorized to access this profile.");
+        }
+
         service.find(id)
                 .ifPresentOrElse(
                     user -> service.update(updateUserWithRequest.apply(user, patchUserRequest)),
@@ -107,36 +126,5 @@ public class UserDefaultController implements UserController {
                     }
                 );
         return getUser(id);
-    }
-
-    public void updateUserPassword(BigInteger id, PutPasswordRequest putPasswordRequest) {
-        service.find(id)
-                .ifPresentOrElse(
-                        user -> service.update(updateUserPasswordWithRequestFunction.apply(user, putPasswordRequest)),
-                        () -> {
-                            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-                        }
-                );
-    }
-
-    @Override
-    public AuthenticationResponse createAuthenticationToken(AuthenticationRequest authenticationRequest, HttpServletResponse response) throws BadCredentialsException, DisabledException, UsernameNotFoundException, IOException {
-        try {
-            authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(authenticationRequest.getEmail(), authenticationRequest.getPassword()));
-        } catch (BadCredentialsException e) {
-            throw new BadCredentialsException("Incorrect email or password");
-        } catch (DisabledException e) {
-            response.sendError(HttpServletResponse.SC_NOT_FOUND, "User not found, register user first");
-            return null;
-        }
-        final UserDetails userDetails = authenticationService.loadUserByUsername(authenticationRequest.getEmail());
-        final String jwt = jwtUtil.generateToken(userDetails.getUsername());
-        final GetUserResponse user = service.find(authenticationRequest.getEmail())
-                .map(userToResponse)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
-        return AuthenticationResponse.builder()
-                .jwt(jwt)
-                .user(user)
-                .build();
     }
 }
